requirements.txt
asana==4.0.10
asttokens==2.4.0
azure-functions==1.17.0
backcall==0.2.0
certifi==2023.7.22
decorator==5.1.1
executing==1.2.0
greenlet==2.0.2
ipython==8.15.0
jedi==0.19.0
matplotlib-inline==0.1.6
numpy==1.25.2
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.20
stack-data==0.6.2
tabulate==0.9.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
